third_party/blink/web_tests/platform/linux-chrome/external/wpt/performance-timeline/not-restored-reasons/performance-navigation-timing-cross-origin-bfcache.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] RemoteContextHelper navigation using BFCache
  assert_equals: expected "no" but got "yes"
Harness: the test ran to completion.
